Revert changing FOR and EXCEPT variable token types. #4708
We had changed the token type used with FOR variables and with
EXCEPT's AS variable from VARIABLE to ASSIGN for them to be consistent
with variables assigned with keyword calls. It later turned out that
we still use VARIABLE in other places for similar purposes so
inconsistency remains. In addition to that, we have noticed some other
inconsistencies and problems with token types. We don't have time to
properly go through them all now in RF 7. Better to postpone this to
RF 8 altogether so that we aren't making backwards incompatible token
type changes in every release.

Also some error message tuning.