Don't add normal keyword or control structure messages to log.
Issue #4883 added `message` to keywords and control structures in the
result model and they were also added to the log file. With nested
structures, it is possible that multiple keywords or control
structures have the same message that is also shared by the test
case. This is somewhat annoying in general and especially problematic
with longer messages.

This commit removes writing normal keyword and control structure
messages to the log file. Only messages that are written are notes
about keywords being removed or flattened with `--remove-keywords` or
`--flatten-keywords`. These notes are now wrapped with `<span
class="robot-note"></span>` to make it possible to recognize them.

Issue #4931 proposes collapsing long messages in log and
report. If/when that's implemented, we can consider including also
keyword and control structure messages in the log file.