Avoid scientific notation with elapsed time in output.xml
It seems that string representation of floats that are smaller than
`0.0001` uses scientific notation which we don't want. For example,
`str(0.00004) == '4e-05'`.

Formatting using the `f` modifier solves the issue. It uses six digits
by default which is fine, but it unfortunately doesn't omit trailing
zeros. They could be rstrippped, but it could strip all digits and
leave us with something like `3.`. Stripping also `.` would work, but
then the result could be just `3` which wouldn't technically amyore be
`xs:float`. Easier to just always use six digits.

Related to #4258.