Handle libraries with attrs having invalid __getattr__.
This fixes a regression in RF 7.0a1 reported with SSHLibrary. It has
some kind of a config object with `__getattr__` that can raise other
exceptions than AttributeError. These exceptions go through when we use
`hasattr(candidate, 'robot_name')` and need to be handled in our code.
I consider SSHLibrary being buggy, but because it worked earlier
this can be consired a regression.