Add KeywordImplementation base class.
It is used now by UserKeyword, will be soon be used by LibraryKeyword,
and may be used by InvalidKeyword or something like that in the
future.  Having a common base class for keyword implementations is
good idea in general, but especially convenient when they will be
exposed to users via listener v3 `start/end_keyword` methods. #3296