Add `KeywordImplementation.params` as an alias for `args`.
When talking about arguments *accepted* by keywords, it is more
correct to use term *parameter* than *argument*. The latter is the
term to use when talking about arguments *passed* to a keyword, but
these terms are often used interchangeably.

Using `params` would thus be better than using `args` in this
context. Changing it isn't worth the effort anymore before RF 7.0, but
having a forward compatible alias makes it easier to change it later.