Enhance TypeInfo.
1. Represent `x` and `x[]` so that they can be separated. Earlier both
   were reprecented with `TypeInfo(name='x', nested=())`, but nowadays
   `nested` is `None` with the first one. This also makes it possible
   to properly represent the first parameter of `Callable[[], None]`.

2. Fix `__str__` if there are nested types but no name. This occurs,
   for example, with `Callable[[], None]` and `Callable[[int], None]`.
   Error reporting gets easier now that `TypeInfo` always have correct
   string representation.

The former change is somewhat backwards incompatible, but becaue
`TypeInfo` only used by Libdoc (and not part of the public API) priort
to RF 7.0, that ought to be ok still after the first release
candidates. All types still have `nested` as a list in Libdoc spec
files for backwards compatibility reasons, but that should be changed
in the future.

`TypeInfo` was made more widely used as part of #4711. Callable
enhancement were inspired by #5002.