PDF version of release not as a workaround
GitHub fails to render rf-7.0.rst file. Here's a PDF generated based
on it that can be viewed until the issue is fixed.

The rendering issue seems to be a problem at GitHub. It cannot anymore
show rf-7.0rc3.rst correctly either and it certainly worked fine
earlier this week when rc3 was created.