Enable ANSI codes on Windows if terminal supports them.
Earlier ANSI codes were always disabled on Windows by default and we
used Windows APIs for setting colors on the console. Now we detect are
ANSI codes supported and use them if they are. The biggest benefit is
that terminal hyperlinks that are implemented using ANSI codes now
work also on Windows if the terminal supports them. #5189