Run WHILE loop once as NOT RUN also if limit is invalid.
PR #5192 (issue #5194) changed WHILE loop behavior so that they
weren't run at all if the limit was invalid. That is inconsistent with
other error situations where the loop is run once as NOT RUN. This
incostency is now fixed and tests updated to validate it. Tests were
enhanced a bit also otherwise.