Add notes about some BuiltIn keywords not supporting named arguments.
Some BuiltIn keywords disable resolving variables in arguments and
that also disables the named-argument syntax. With most of these
keywords that doesn't matter, but with the following three using the
`default` or `message` argument like `default=value` would be logical:

- Get Variable Value
- Variable Should Exist
- Variable Should Not Exist

It would be better to cleanup the messy system for disabling variables
from being resolved, but that's too much work at the moment. This
commit adds notes about the limitation to the affected keywords.

Fixes #5235.