Initiallize `TypeConverter.languages` only when needed.
`TypeConverters` gets `languages` that can be a `Languages` instance
or `None`. In the latter case a new `Languages` instance is needed if
someone accesses `TypeConverters.languages`. Earlier `Languages` was
created already in `__init__`, but now it is done only if
`TypeConverter.languages` is actually accessed.

This doesn't affect execution at all, because then `langauges` is
always a `Languages` instance. With Libdoc `languages` is always
`None` and `TypeConverters.languages` is never accessed, so lazy
initialization saves time. Apparently initializing `Languages` is
rather slow, because this causes roughly 50% performance enhancement
with really big libraries. Investigating why initialization is so slow
could be a good idea, but this fix makes sense anyway.

Fixes #5254.