Add `keywords`, `messages` and `non_messages` to test result model.
This avoid the need to use `body.filter(keywords/messages=True/False)`
in tests. `robot.result.Keyword` already has `message`, but other body
items don't.

This is related to the removal `kws` and `msgs` that was done in the
previous commit. In most cases accessing `body` is fine and removing
the old aliases was good in general, but the somewhat complex
`filter()` approach was needed in few places.