Handle logging and keyword running listeners with JSON output.
Listeners logging or running keywords in strange places cause
unexpected data to be logged. Make sure such data is handled
correctly also when using JSON output (#3423).

The exact output isn't tested as thoroughly as with XML. It is more
important to make sure that tests are executed as expected and outputs
aren't corrupted.