Debugfile available for threads/processes/async
This is possible by using a single process which accepts write requests
via a queue.

This lock and thread free design brings the following advantages over threads.:

- threads, fork, async, subprocess and in the future also subinterpreter based
parallelism is possible.

As soon as a single thread is started, Fork and fork based subprocesses
are no longer an option.

Whole classes of bugs are not aplicable
- life lock
- dead lock

Outlook.:
subinterpreter and subprocesses on windows are on the todo list.