Process: Support Robot's timeouts also on Windows.
The fix is based on the code in PR #5302 by @franzhaas.

Fixes #5345.

Now that `Popen.communicate()` gets a timeout, its implementation
always gets to a code path where stdtout/stderr being a closed PIPE
doesn't matter and we only need to care about stdin. As the result
content written to PIPEs that are later closed is now available.