Add application and taskbar icons to Dialogs.
It depends on OS how/where icons are shown. For example, on Linux with
Gnone there's no icon in the dialog, but there's a taskbar icon as
well as an icon in the application switcher. On Windows there's a
taskbar icon and also the dialog itself has an icon. OSX is yet to be
tested.

This is part of #5334.