Process: Don't log timeout type because its unreliable.
On Windows the timeout logic raises a `TimeoutError` without any
parameters and thus its `kind` is always `TEST`. That means we cannot
reliaby log the type of the occurred timeout.

`TimeoutError.kind` was added for exactly this purpose in 2b9f0c7.
Better to remove those changes and also document that existing
`test/keyword_timeout` attributes aren't part of the public API.

Fixes a bug in the implementation of #5376.