Dialogs: Activate default selection.
When `Get Selection From User` is used with a default value, not only
select that value but also activate it. That affects moving the
selection with arrow keys. Now the initial position is the selection,
earlier it was the first item.

To some extend related to Dialog look and feel enhancements
(#5334). Too small fix to deserve its own issue.