Fix for delaying logging with timeouts.
Previous messages weren't restored meaning that messages could be
lost. The unused variable also made linters unhappy.

This code attemps to avoid problems with timeouts interrupting writing
to output files. The current code was written to fix #5395 and is
basically a rewrite of the fix for #2839. As #5417 explains, there are
still problems and bigger changes are needed.