Unit tests for pausing timeouts.
Also small adjustment to when to raise a timeout on resume. Earlier
timeout was raised on the first resume after timeout had been
exceeded, but now the runner must be fully resumed. This behavior is
more logical than the earlier.

The change shouldn't affect execution at all. The reason is that if
timeout is paused in nested manner, `BuiltIn.run_keyword` has been
used in recursively and each recursive call has its own timeout. The
last one of them will raise a timeout on resume immediately after the
timeout has been exceeded anyway.

This is part of #5417.