`robot.utils._secret` -> `robot.utils.secret`
We'd like the `Secret` type (#4537) to be implemented under `robot.api`,
but that causes very-hard-to-resolve cyclic import issues. After some
failed attempts, we decided to keep it under `robot.utils`.

At some point the idea was to "hide" the implementation under `robot.utils`
by naming the module `_secret`, but we haven't used such convention with
other modules so better to go with normal `secret` name. The docs will
be updated later to explain that the class should be imported under
`robot.api.types`.