Support not printing errors and warnings to console (#5462)
There are use cases for printing a warning or error message to the log
but not to the console.

This commit adds an optional also_console argument to logger.write,
logger.warn, and logger.error. Its value is stored in the message object
and based on it the loggers in src/robot/output/console/ print the
message or not.

Fixes #5460.

* atest: Update "Write messages to console" test

Messages logged with `also_console=True` are now printed to stderr
instead of stdout (see previous commit).

* atest: Test writing warning and error messages only to log