Make timeout tests faster and hopefully more stable
Shorter timeouts and sleeping times. This kind of changes are a bit
risky, because they can make tests flakey. In this case changes should
make tests more stable, though, because the default test timeout has
been increased. Even if a test would be slow on some run (e.g. due to
OS), it should not hit the new 10s timeout.